requirements.txt
azure-functions
azure-storage-blob6.0
azure-cosmos
azure-identity
azure-keyvault-secrets
psycopg2-binary0
google-genai
langchain
langchain-community
pypdf
requests